Split video_analyzer.py into focused modules
The 2,500-line video_analyzer.py mixed prompt definitions, ffmpeg
helpers, Gemini transport, retry logic, sanity checks, and response
shaping in a single module. Targeted edits were high-risk because
scanning for a single constant meant reading through unrelated
concerns.

Split into api/src/wcs_api/services/video_analysis/ with five
focused modules:

- prompts.py          — PATTERN_SEGMENTATION_PROMPT,
                        GEMINI_VIDEO_PROMPT, SYSTEM_PROMPT,
                        _build_user_context,
                        _build_sanity_retry_prompt,
                        _format_pattern_timeline
- media_context.py    — _extract_beat_context,
                        _detect_motion_floor, ffprobe helpers,
                        Beat This! wrappers, _video_part_with_fps,
                        get_video_duration
- gemini_client.py    — _call_gemini, _build_gen_config,
                        _safe_parse_json, upload+poll helpers
- response_sanitizer.py — _sanitize_patterns,
                        _sanitize_musical_moments,
                        _sanitize_follower_initiative,
                        _sanity_check, _pattern_inside_window,
                        _safe_float, _summarize_patterns,
                        _coerce_score, _coerce_str_list,
                        _coerce_category, _shape_response,
                        _compute_overall, _grade_letter,
                        _max_interval, _sanitize_user_field,
                        name/variant normalization
- analyzer.py         — analyze_video_path, _run_pattern_pre_pass,
                        analyze_video_bytes, VideoAnalysisError

video_analyzer.py is now a one-line shim
(from .video_analysis import *) so routes/video.py and every
existing 'from ..services.video_analyzer import ...' caller keeps
working unchanged.

No behavior change: verified via smoke tests that analyze_video_path
resolves to the same function through both import paths, that the
newly added coerce helpers still validate malformed inputs
correctly, that _build_sanity_retry_prompt accepts dance_start_sec
/ dance_end_sec kwargs, and that _format_pattern_timeline still
renders variant + visual_cue.

Note: gemini_client.py carries two helpers (_upload_video_file,
_delete_uploaded_file) that analyze_video_path doesn't call — the
facade still inlines its upload / poll / cleanup flow. Left them
in place because removing them was out of scope for the move, but
they're fair game for a follow-up cleanup.

Closes #79.
